[REFACTOR] Task handler설정
