+++ src/main/java/com/uplog/uplog/domain/task/application/TaskService.java
@@ -7,7 +7,7 @@ import com.uplog.uplog.domain.menu.model.Menu;
 import com.uplog.uplog.domain.task.dao.TaskRepository;
 import com.uplog.uplog.domain.task.dto.TaskDTO.*;
 import com.uplog.uplog.domain.task.model.Task;
-import com.uplog.uplog.domain.task.exception.handler.*;
+import com.uplog.uplog.domain.task.exception.*;
 import com.uplog.uplog.domain.task.model.TaskStatus;
 import com.uplog.uplog.domain.team.dao.ProjectTeamRepository;
 import com.uplog.uplog.domain.team.model.ProjectTeam;
